git-svn-id: https://svn.code.sf.net/p/openpfgw/code@396 368e28a5-49e5-40c5-be6c-d84989c25d7b

+++ pform/pfgwlib/gwcontext.cpp
@@ -51,8 +51,8 @@ int CreateModulus(Integer *NN, bool kbncdEligible, int increaseFFTSize)
    char  testString[100];
    int   error_code;
    double   k;
-   unsigned long b, n;
-   long c, d;
+   uint32 b, n;
+   int32 c, d;
 
    sprintf(testString, "%send1", g_cpTestString);
